LibWeb/CSS: Revert to PercentageOr::resolve() to be templated
In dd3a0afbdb77b09c212182d2329d39fac51d7245 this function was modified
to accept only Angle instead of being templated, with the purpose of
preventing Length from being passed.

This change reverts the function to being templated but adds a
constraint specifically forbidding Length. This adjustment will allow
for the future use of Frequency, Number, Time, etc., if we would want.