LibWeb: Empty CE reaction queue instead of destroying it on exception
If an exception occurs in a custom element constructor, we clear the
reaction queue by destroying it, instead of emptying the Vector.
https://github.com/SerenityOS/serenity/blob/9c2ec14745ad6702e7fb5c02d454696b9509a6fa/Userland/Libraries/LibWeb/DOM/Element.cpp#L2033

This causes a UAF here, as async upgrades (i.e. custom elements not
created by document.createElement) are performed in this loop:
https://github.com/SerenityOS/serenity/blob/9c2ec14745ad6702e7fb5c02d454696b9509a6fa/Userland/Libraries/LibWeb/Bindings/MainThreadVM.cpp#L657

Fixes crash when loading https://github.com/SerenityOS/serenity